requirements.txt
attrs==22.1.0
black==22.10.0
click==8.1.3
exceptiongroup==1.0.1
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-magic==0.4.27
tomli==2.0.1
